app/requirements.txt
uvicorn==0.13.1
gunicorn==20.0.4
Django==3.0.6
psycopg2==2.8.5
djangorestframework==3.11.0
drf-yasg==1.17.1
uvloop==0.14.0
httptools
pylint